+++ packages/@sanity/vanilla-extract-tsdown-plugin/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sanity/vanilla-extract-tsdown-plugin
 
+## 0.3.1
+
+### Patch Changes
+
+- [#3214](https://github.com/sanity-io/pkg-utils/pull/3214) [`23788f8`](https://github.com/sanity-io/pkg-utils/commit/23788f85919c8d4b511110e5f93494d41506a4e0) Thanks [@squiggler-app](https://github.com/apps/squiggler-app)! - fix(deps): update dependency rolldown to ~1.2.3
+
+- Updated dependencies [[`23788f8`](https://github.com/sanity-io/pkg-utils/commit/23788f85919c8d4b511110e5f93494d41506a4e0)]:
+  - @sanity/vanilla-extract-rolldown-plugin@0.4.1
+
 ## 0.3.0
 
 ### Minor Changes